// resources/js/types/enums.ts
export enum UserRole {
    USER = 'user',
    ADMIN = 'admin',
}

export enum IdeaStatus {
    PENDING = 'pending',
    APPROVED = 'approved',
    REJECTED = 'rejected',
    WINNER = 'winner',
}

export enum PrizeStatus {
    PENDING = 'pending',
    DELIVERED = 'delivered',
}

export enum SponsorshipStatus {
    PENDING = 'pending',
    APPROVED = 'approved',
    REJECTED = 'rejected',
}